Act3 i Act7 optimitzats per memoria

+++ CODE INSITUT CFGS/Programacio/Java Acabat/UD2/UD2Act3MarcMas/src/main/java/org/yourcompany/yourproject/UD2Act3MarcMas.java
@@ -15,7 +15,7 @@ public class UD2Act3MarcMas {
         // Activitat 3
         Scanner s = new Scanner(System.in);
         System.out.println("Introdueixi un numero");
-        Double n1 = s.nextDouble(); // Llegesc input d'usuari
+        Float n1 = s.nextFloat(); // Llegesc input d'usuari
         if (n1 < 1 && n1 > -1 && n1 != 0) { //si el nombre es major a -1, menor que 1, i no es 0 és quasi-zero
             System.out.println("El numero es quasi-zero");
         } else {
